+++ CHANGELOG.md
@@ -5,3 +5,8 @@ This for documenting every version, most important details will be README.md but
 ## [[0.0.1]]
 
 Initial version, no gameplay, just some basic variables and basic code
+
+## [[0.0.2]]
+
+sprites can now move if you make it table and add timer etc.
+Framework for better sprite work added